// src/types/generated/patientDetailsDto.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * Doctaura API
 * API for managing clinic operations, doctor schedules, and appointments
 * OpenAPI spec version: v1
 */
import type { Gender } from './gender';

export interface PatientDetailsDto {
  id?: string;
  /** @nullable */
  userId?: string | null;
  /** @nullable */
  firstName?: string | null;
  /** @nullable */
  lastName?: string | null;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  email?: string | null;
  dateOfBirth?: string;
  gender?: Gender;
  /** @nullable */
  countryId?: number | null;
  /** @nullable */
  governorateId?: number | null;
  /** @nullable */
  districtId?: number | null;
  /** @nullable */
  localityId?: number | null;
  /** @nullable */
  emergencyContactName?: string | null;
  /** @nullable */
  emergencyContactPhone?: string | null;
  /** @nullable */
  bloodType?: string | null;
  /** @nullable */
  allergies?: string | null;
  /** @nullable */
  chronicConditions?: string | null;
  /** @nullable */
  preferredLanguage?: string | null;
  isActive?: boolean;
  createdAt?: string;
  updatedAt?: string;
}